Updated client and API
Added a language selection on account creation so when the user has logged in they will automatically see the quizes relevant to them.
Added additional premade quiz exampls for other languages.
Updated account creation to require the password to be entered twice.
Updated API to correctly delete a user and all the notes they have stored.